New digAPI
Created digAPI and changed stripTunnel and mineTunnel to use these.
